+++ Sources/TuistCache/Cache/CloudHEADResponse.swift
@@ -1,0 +1,13 @@
+import Foundation
+import TuistCore
+import TuistSupport
+
+struct CloudHEADResponse: Decodable {
+    public static func existsResource(hash: String, config: Config) throws -> HTTPResource<CloudResponse<CloudHEADResponse>, CloudResponseError> {
+        guard let configCloud = config.cloud else { throw CacheAPIError.missingCloudConfig }
+        let url = try URL.apiCacheURL(hash: hash, cloudURL: configCloud.url, projectId: configCloud.projectId)
+        var request = URLRequest(url: url)
+        request.httpMethod = "HEAD"
+        return .jsonResource { request }
+    }
+}