+++ Sources/TuistGeneratorTesting/Linter/MockGraphLinter.swift
@@ -1,6 +1,5 @@
 import Foundation
 import TuistCore
-import TuistCoreTesting
 import TuistSupport
 @testable import TuistGenerator