+++ features/step_definitions/build.rb
@@ -1,8 +1,12 @@
 Then(/^tuist builds the project$/) do
   system("swift", "run", "tuist", "build", "--path", @dir)
 end
-Then(/^tuist builds the scheme (.+) from the project$/) do |scheme|
+Then(/^tuist builds the scheme ([a-zA-Z]+) from the project$/) do |scheme|
   system("swift", "run", "tuist", "build", scheme, "--path", @dir)
+end
+
+Then(/^tuist builds the scheme ([a-zA-Z]+) and configuration ([a-zA-Z]+) from the project$/) do |scheme, configuration|
+  system("swift", "run", "tuist", "build", scheme, "--path", @dir, "--configuration", configuration)
 end
 
 Then(/^tuist builds the project at (.+)$/) do |path|